Saving System Functional
Selected Meshes (In object mode) can now be saved into a .json file as a dictionary listing all the verts required to create the faces and all of the locations of each vert.

Doing it edge by edge was not functioning so all meshes now need to be modelled with faces and saved that way - Removing the faces is done easilly when the objects are "loaded" aka, the .json is read and the library stored inside is used to create a new object using blenders base system -- This can be improved by utilising the bmesh system to avoid adding too much orphan data to a scene; however this proved to be massively outside of the scope of this project.